+++ lib/components/AppBar/AppBarTitle.d.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+export interface IAppBarTitleProps {
+    children?: ReactNode;
+    onClick?: () => void;
+}
+export declare const AppBarTitle: ({ children, onClick }: IAppBarTitleProps) => JSX.Element;